chore: codegen

+++ frb_example/dart_minimal/lib/src/rust/api/minimal.dart
@@ -9,5 +9,5 @@ import 'package:flutter_rust_bridge/flutter_rust_bridge_for_generated.dart';
 Future<int> minimalAdder({required int a, required int b}) =>
     RustLib.instance.api.crateApiMinimalMinimalAdder(a: a, b: b);
 
-Future<BigInt> f({required BigInt arg}) =>
+Future<int> f({required int arg}) =>
     RustLib.instance.api.crateApiMinimalF(arg: arg);
